+++ src/manager/index.ts
@@ -1,4 +1,7 @@
 export {Request, RequestFactory} from './http';
-export {ManagerOptions,ManagerBase} from './base';
+export {ModelManager} from './manager';
 
-export {Search, SearchParameter, SearchParameterMap, SearchResultPage, SearchResult} from './search';
+export {
+    Search, SearchParameter, SearchParameterMap, SearchResultPage, SearchResult,
+    SEARCH_PAGE_SIZE, SEARCH_PAGE_QUERY_PARAM
+} from './search';